% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefasRBR-package.R
\docType{package}
\name{cefasRBR-package}
\alias{cefasRBR}
\alias{cefasRBR-package}
\title{cefasRBR: R interface to RBR RSK files}
\description{
Functions for working with RBR profiler data, e.g. Concerto and Maestro instruments.
}
\keyword{internal}
