% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbr_channels}
\alias{rbr_channels}
\title{RBR logger3 command set - channel types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 158 rows and 5 columns.
}
\source{
<https://docs.rbr-global.com/display/L3commandreference/Supported-Channel-Types>
}
\usage{
rbr_channels
}
\description{
A list of the supported channel types by the RBR logger3 firmware, and a description of what those channels are.
Used by `read.rsk`
}
\keyword{datasets}
