% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbr_error_codes}
\alias{rbr_error_codes}
\title{RBR logger3 command set - error codes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 26 rows and 2 columns.
}
\source{
<https://docs.rbr-global.com/L3commandreference/format-of-stored-data/easyparse-calbin00-format/sample-data-easyparse-format>
}
\usage{
rbr_error_codes
}
\description{
A list of the supported error codes by the RBR logger3 firmware, and a description of what those errors are.
Used by `read.rsk`
}
\keyword{datasets}
