% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbr_event_codes}
\alias{rbr_event_codes}
\title{RBR logger3 command set - event codes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 42 rows and 2 columns.
}
\source{
<https://docs.rbr-global.com/L3commandreference/format-of-stored-data/easyparse-calbin00-format/easyparse-format-events-markers>
}
\usage{
rbr_event_codes
}
\description{
A list of the supported event codes by the RBR logger3 firmware, and a description of what those events are.
Used by `read.rsk`
}
\keyword{datasets}
