% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk.R
\name{rsk.activations}
\alias{rsk.activations}
\title{Process the activation times from a RBR instrument}
\usage{
rsk.activations(rsk, min_length_activation = 120)
}
\arguments{
\item{rsk}{an object of class cefasRSK}

\item{min_length_activation}{how short a activation can be in seconds}
}
\value{
cefasRSK object
}
\description{
Process the activation times from a RBR instrument
}
