% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk.R
\name{rsk.addgeoregion}
\alias{rsk.addgeoregion}
\title{Insert location metadata into .RSK file from table}
\usage{
rsk.addgeoregion(filename, tbl)
}
\arguments{
\item{filename}{.rsk file path}

\item{tbl}{a data.frame containing the following columns
dateTime - POSIXct timestamp
latitude - in decimal degrees
longitude - in decimal degrees
label - a station code, can be numeric or text
description - text, can be empty}
}
\value{
nothing
}
\description{
Insert location metadata into .RSK file from table
}
\examples{
arb_data = data.table(
label = c(10, 55),
description = c("arbitary position data"),
dateTime = as.POSIXct(c("2024-04-21 08:19", "2024-04-21 09:30"), tz = "UTC"),
latitude = c(52.1, 52.12),
longitude = c(2, 2.1)
)
## Not run:
rsk.addgeoregion("myrsk.rsk", arb_data)
## End(Not run)
}
