% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk.R
\name{rsk.recalculate_profiles}
\alias{rsk.recalculate_profiles}
\title{Recalculate profiles for a .RSK file}
\usage{
rsk.recalculate_profiles(filename, salinity_tol = 0.1, prs_tol = 0.1)
}
\arguments{
\item{filename}{the .RSK file}

\item{salinity_tol}{the minimum salinity for the start of a profile, default = 0.1}

\item{prs_tol}{the threshold increase in pressure over the minimum, default is 0.1 (i.e. minimum recorded pressure + 0.1 hPa)}
}
\value{
writes directly to the .RSK
}
\description{
This tool gives you a little more control on the automated profile detection provided by ruskin
}
\details{
Note: this tool will overwrite any annotations already present in the .RSK, use with caution
 Note: this tool does not attempt to identify casts, TODO
}
