% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk.R
\name{rsk.regions}
\alias{rsk.regions}
\title{Add profile label and GPS metadata to extracted data}
\usage{
rsk.regions(rsk)
}
\arguments{
\item{rsk}{an object of class cefasRSK}
}
\value{
cefasRSK object
}
\description{
Add profile label and GPS metadata to extracted data
}
